+++ ios/PacketTunnel/vpncloud-Bridging-Header.h
@@ -1,0 +1,6 @@
+#include <string.h>
+#include <sys/ioctl.h>
+#include <sys/kern_control.h>
+#include <sys/socket.h>
+#include <sys/sys_domain.h>
+#include "vpncloud.h"