+++ ios/VpnCloud/VpnCloudApp.swift
@@ -1,0 +1,13 @@
+import SwiftUI
+
+@main
+struct VpnCloudApp: App {
+    @StateObject private var vpn = VpnManager()
+
+    var body: some Scene {
+        WindowGroup {
+            ContentView()
+                .environmentObject(vpn)
+        }
+    }
+}