+++ src/lib.rs
@@ -41,6 +41,9 @@ pub mod wsproxy;
 #[cfg(target_os = "android")]
 mod android;
 
+#[cfg(target_os = "ios")]
+mod ios;
+
 #[cfg(windows)]
 pub mod tray;